alpha/backend/requirements.txt
python-dotenv==1.1.1
ibm-watson==10.0.0
ibm_watsonx_ai==1.3.32
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20